Refactor: Move SplashRoute and improve code style
