+++ presentation/src/main/java/com/london/presentation/screen/NetworkErrorScreen.kt
@@ -21,7 +21,7 @@ fun NetworkErrorScreen() {
     Column(
         modifier = Modifier
             .fillMaxSize()
-            .padding(horizontal = 16.dp), // Added horizontal padding
+            .padding(horizontal = 16.dp),
         verticalArrangement = Arrangement.Center,
         horizontalAlignment = Alignment.CenterHorizontally
     ) {
