+++ presentation/src/main/java/com/london/presentation/screen/home/HomeScreenUiState.kt
@@ -14,6 +14,6 @@ data class HomeScreenUiState(
     val error: ErrorState? = null,
     val popularMovies: List<PopularMovie> = emptyList(),
     val popularTvShows: List<PopularTvShow> = emptyList(),
-    val upcomingMovies:  StateFlow<PagingData<Movie>> = MutableStateFlow<PagingData<Movie>>(PagingData.empty()),
+    val upcomingMovies: StateFlow<PagingData<Movie>> = MutableStateFlow(PagingData.empty()),
     val selectedGenre: MovieGenre = MovieGenre.All
 )
