+++ presentation/src/main/java/com/london/presentation/screen/login/LoginContract.kt
@@ -3,13 +3,13 @@ package com.london.presentation.screen.login
 import androidx.compose.ui.text.input.TextFieldValue
 
 interface LoginContract {
- fun onUsernameChanged(username: TextFieldValue)
- fun onPasswordChanged(password: TextFieldValue)
- fun onPasswordVisibilityToggled()
- fun onLoginClick()
- fun onLoginAsGuestClick()
+    fun onUsernameChanged(username: TextFieldValue)
+    fun onPasswordChanged(password: TextFieldValue)
+    fun onPasswordVisibilityToggled()
+    fun onLoginClick()
+    fun onLoginAsGuestClick()
 
- fun onNavigateBack()
- fun onCreateAccountClick()
- fun onForgotPasswordClick()
+    fun onNavigateBack()
+    fun onCreateAccountClick()
+    fun onForgotPasswordClick()
 }
